COMP0016_2020_21_Team12-datasetsExperimentsAna/Main/src/EventGridTriggerAzureFunction/requirements.txt
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
azure-cosmosdb-table
azure-storage-blob
azure-storage-queue
azure-data-tables
grpcio==1.14.2
grpcio-tools==1.14.2
protobuf==3.6.1
six==1.11.0